Implements delete review and delete listing functionality, restructures routes using Express Router and updates UI styling
