+++ app.js
@@ -3,12 +3,11 @@ const app = express();
 const path = require("path");
 const methodOverride = require("method-override");
 const mongoose = require("mongoose");
-const Listing = require("./models/listing.js");
-const wrapAsync = require("./utils/wrapAsync.js");
 const ExpressError = require("./utils/ExpressError.js");
-const { listingSchema, reviewSchema } = require("./schema.js");
 const ejsMate = require("ejs-mate"); 
-const Review = require("./models/review.js")
+
+const listings = require("./routes/listing.js");
+const review = require("./routes/review.js");
 
 const MONGO_URL = "mongodb://127.0.0.1:27017/staypoint";
 main()
@@ -31,102 +30,17 @@ app.get("/", (req, res) => {
     res.send("root is working!");
 });
 
-const validateListing = (req, res, next) => {
-    let {error} = listingSchema.validate(req.body);
-    if(error){
-        let errMsg = error.details.map((el) => el.message).join(",");
-        throw new ExpressError(400, errMsg);
-    }
-    else{
-        next();
-    }
-}
-
-const validateReview = (req, res, next) => {
-    let {error} = reviewSchema.validate(req.body);
-    if(error){
-        let errMsg = error.details.map((el) => el.message).join(",");
-        throw new ExpressError(400, errMsg);
-    }
-    else{
-        next();
-    }
-}
-
-//  Index route
-app.get("/listings", wrapAsync(async (req, res) => {
-    const allListings = await Listing.find({});
-    res.render("listings/index.ejs", {allListings});
-}));
-
-//  New route
-app.get("/listings/new", (req, res) => {
-    res.render("listings/new.ejs");
-});
-
-//  Show route
-app.get("/listings/:id", wrapAsync(async (req, res) => {
-    let {id} = req.params;
-    const listing = await Listing.findById(id);
-    res.render("listings/show.ejs", {listing});
-}));
-
-//  Create route
-app.post("/listings", validateListing, wrapAsync(async (req, res, next) => {
-    // if(!req.body.listing){
-    //     throw new ExpressError(400, "Send valid data for listing");
-    // }
-    const newListing = new Listing(req.body.listing);
-    await newListing.save();
-    res.redirect("/listings"); 
-}));
-
-//  Edit route
-app.get("/listings/:id/edit", wrapAsync(async (req, res) => {
-    let {id} = req.params;
-    const listing = await Listing.findById(id);
-    res.render("listings/edit.ejs", {listing});
-}));
-
-//  Update route
-app.put("/listings/:id", validateListing, wrapAsync(async (req, res) => {
-    let {id} = req.params;
-    await Listing.findByIdAndUpdate(id, {...req.body.listing});
-    res.redirect(`/listings/${id}`);
-}));
-
-//  Delete route
-app.delete("/listings/:id", wrapAsync(async (req, res) => {
-    let {id} = req.params;
-    let deletedListing = await Listing.findByIdAndDelete(id);
-    console.log(deletedListing);
-    res.redirect("/listings");
-}));
-
-// Reviews
-// Post route
-app.post("/listings/:id/reviews", validateReview, wrapAsync(async (req, res) => {
-    let listing = await Listing.findById(req.params.id);
-    let newReview = new Review(req.body.review); 
-
-    listing.reviews.push(newReview);
-
-    await newReview.save();
-    await listing.save();
-    
-    res.redirect(`/listings/${listing._id}`);
-}))
-
+app.use("/listings", listings);
+app.use("/listings/:id/reviews", review);
 
 // app.all("*", (req, res, next) => {
-//     next(new ExpressError(404, "Page not found!"));
+//     next(new ExpressError(404, "Page not found"));
 // });
 
 // error handling middleware
 app.use((err, req, res, next) => {
     let{statusCode = 500, message = "Something went wrong!"} = err;
     res.status(statusCode).render("error.ejs", {message});
-    // res.status(statusCode).send(message);
 });
 
 app.listen(8080, () => {
