+++ models/listing.js
@@ -1,5 +1,6 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
+const Review = require("./review.js");
 
 
 const listingSchema = new Schema({
@@ -36,6 +37,12 @@ const listingSchema = new Schema({
             ref: "Review"
         }
     ]
+});
+
+listingSchema.post("findOneAndDelete", async (listing) => {
+    if(listing){
+        await Review.deleteMany({_id: {$in: listing.reviews}});
+    }
 });
 
 const Listing = mongoose.model("Listing", listingSchema);
