+++ public/css/style.css
@@ -51,7 +51,7 @@ body{
 }
 
 /* Cards */
-.card{
+.listing-card{
     border: none !important;
     margin-bottom: 2rem;
     transition: transform 0.3s ease;
