+++ views/listings/index.ejs
@@ -8,7 +8,7 @@
     <div class="row row-cols-lg-3 row-cols-md-2 row-cols-sm-1">
         <% for(let listing of allListings){ %>
             <a href="/listings/<%= listing._id %>" class="listing-link">
-                <div class="card col">
+                <div class="card col listing-card">
                     <img src="<%= listing.image %>" class="card-img-top" alt="lisgting_image" style="height: 20rem;">
                     <div class="card-body">
                         <p class="card-text">
