+++ views/listings/show.ejs
@@ -3,7 +3,7 @@
     <div class="col-8 offset-2">
         <h3> <%= listing.title %> </h3> <br>
     </div>
-    <div class="col-6 offset-2 show-card">
+    <div class="card col-6 offset-2 show-card listing-card">
         <img src="<%= listing.image %>" class="card-img-top show-image" alt="listing_image"> <br><br> 
         <div class="card-body col-6">
             <p class="card-text">
@@ -40,5 +40,23 @@
             </div>
             <button class="btn btn-outline-dark">Submit</button>
         </form>
-    </div>
+        <hr>
+        <p><b>All Reviews</b></p>
+        <div class="row">
+            <% for(review of listing.reviews) { %>
+                <div class="card col-5 ms-3 mb-2">
+                    <div class="card-body">
+                        <h5 class="card-title">Alok Chaturvedi</h5>
+                        <p class="card-text"><%= review.comment %></p>
+                        <p class="card-text"><%= review.rating %> &#127775;</p>
+                        <form class="mb-2 mt-1" method="post" action="/listings/<%= listing._id %>/reviews/<%= review._id %>?_method=DELETE">
+                            <button class="btn btn-sm btn-dark">Delete</button> 
+                        </form>
+                    </div>
+                </div>
+            <% } %>
+        </div>
+    </div> 
+
+
 </div>
